requirements.txt
openai>=1.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
python-dotenv>=1.0.0
typing-extensions>=4.0.0